Rename pkg/procmgr/telemetry to pkg/procmgr/coat
Align with trace COAT naming: the package reports supervision state via
COAT gauges, not telemetry infrastructure. Rename telemetry.go to snapshot.go
for domain types.
