+++ pkg/procmgr/coat/collector_linux.go
@@ -5,7 +5,7 @@
 
 //go:build linux
 
-package telemetry
+package coat
 
 import (
 	"context"
