+++ pkg/procmgr/coat/collector_test.go
@@ -5,7 +5,7 @@
 
 //go:build linux
 
-package telemetry
+package coat
 
 import (
 	"context"
