+++ pkg/procmgr/coat/paths_linux.go
@@ -5,7 +5,7 @@
 
 //go:build linux
 
-package telemetry
+package coat
 
 import (
 	"path/filepath"
