+++ pkg/procmgr/coat/paths_other.go
@@ -5,7 +5,7 @@
 
 //go:build !linux && !windows
 
-package telemetry
+package coat
 
 import (
 	"path/filepath"
